Implement separate rate limit for message submission tasks (#1747)
This adds separate rate limits for tasks depending on their types.
Specifically it implements a tough restriction on the rate of assistant
replies and a fairly strict one on prompter replies.

The logic here is that users should not be spamming many replies in a
short period of time as this is likely impossible without compromising
quality.