Adding Home Page Unit Test (#1858)
Noticed there wasn't a jest test for the home/index page. Decided to add
one in case something caused that to not work and came up as a breaking
change.

Also, I'm planning to add some more website unit tests over the next
week or two. It wouldn't hurt for us to have those.