Add refresh tokens to inference server auth (#2129)
Close #2124.

We also need to have a means of revoking the refresh token from the
backend for security. This will likely mean hashing the refresh token
and storing the hash in a DB table with the user ID and an `enabled`
flag which we can set to false to revoke the token. This can be handled
in a follow-up issue.

This changes some of the existing auth endpoint responses to return both
a refresh token and an access token instead of just an access token, so
will require changes to the website.