Chat config UI (#2115)
There is a small bug in the chat settings button that doesn't display
correctly on mobile. This is due to our layout which is really buggy
lol. I'm going to fix it in the later PR. Also, this PR doesn't store
the config in local storage yet.

**UPDATE**: I added preset, which will allow using pre-defined
parameters in the server. Users could also select a custom preset to
customize the paras as they want. I also remove the params translation
as they really hard to correctly translate. We can add a tooltip for
long descriptions like in the playground of Open in the future.
Current params config:
| name               | min | max  |   |
|--------------------|-----|------|---|
| temperature        | 0   | 1.5  |   |
| max_new_tokens     | 1   | 1024 |   |
| top_p              | 0   | 1    |   |
| repetition_penalty | 1   | 3    |   |
| top_k | 0 | 1 | exponential function: `Math.round(Math.pow(32000, x))`
|

UI

![image](https://user-images.githubusercontent.com/33456881/226656639-2e192759-1fd3-40c2-bb81-aa963519b2f8.png)
Custom preset

![image](https://user-images.githubusercontent.com/33456881/226657164-b80c9794-c7c0-4863-b176-cb7830d2cd13.png)