Show model config in the chat UI (#2317)
So that user knows which config is shown, also moved the config button
next to the submit button.

The PR is more about functionality and less about design. We can change
the design entirely in the future.


![image](https://user-images.githubusercontent.com/24505302/229601629-b4ca4404-fd6a-439d-8650-00251a8faf76.png)


Also, I updated charka UI, since VSCode was having many problems with
it.