Use SSR for every pages (#2334)
~~It's ready now. `_document` only runs once serverside, so this is the
expected behavior we want. `getInitialProps` run on both server and
client. On the client it populates the env wrong, it uses static env
from `.env` instead of actual env, I'm not sure why.~~

Using SSR is the only way to make it work. This PR is quite similar to
#2343 but I expose it to a global var for flexibility (accessing the env
outside of react scope) and more performance.