Implement new chat UI (#2436)
Most of the changes are just UI code. However, we now create chat only
when a user enters a message (this help prevent empty chat). There is a
small flash when creating new chat, I will try to address this in the
next PR.
The current UI have a lot of space between elements. I'm going to do
more breaking change to the chat UI, so we can have a right sidebar
instead of the drawer and we can get rid of the chat config summary, the
new UI will have layout look like Discord.
Desktop UI

![image](https://user-images.githubusercontent.com/33456881/230999113-6434899b-a30d-4eb5-b63d-68f746446421.png)Mobile
UI

![image](https://user-images.githubusercontent.com/33456881/230985336-425603a0-ca8b-4221-8154-7575db9de293.png)
Mobile UI when expand chat list

![image](https://user-images.githubusercontent.com/33456881/230985372-93d289f0-4286-4eba-a6b5-0874a9f82118.png)

---------

Co-authored-by: AbdBarho <ka70911@gmail.com>