Forever Dreaming TV Transcript Dataset (#2631)
I have finally finished crawling foreverdreaming.com. The transcripts
are about 5% of all the content on the website.
However, we have decided not to share the crawler notebook this time,
because it would allow anyone to mirror all the contents on the website
just by changing a few lines of code. The owner of foreverdreaming has
invested a lot of time and resources into running their website and it
simply would not be fair towards them.
We have discussed this on Discord. 

The dataset is https://huggingface.co/datasets/sedthh/fd_dialogue