Instructional Dataset for Stable Diffusion prompt generation (#2982)
I am adding this dataset:

https://huggingface.co/datasets/MadVoyager/stable_diffusion_instructional_dataset

That is an instructional version of this dataset:
https://huggingface.co/datasets/Gustavosta/Stable-Diffusion-Prompts

I believe this should be able to help Open Assistant's ability in
providing accurate Stable Diffusion prompts with image quality tags,
artists and other styling specifications. The Dataset already uses OA's
format, so no script is needed.