Fix Report model field name (#3052)
Close #3051. This turned out to be an easy fix. I don't think the
frontend currently reads this at all so there shouldn't be a
corresponding change to be made there, but correct me if wrong.

I also made a formatting change as pre-commit on commit was conflicting
with `pre-commit run --all-files` (how?)