Plugins fixes (#3070)
This PR addresses the following fixes and modifications:
- Some plugin url-s are with text/json content_type so downloading and
parsing fails. This is now fixed with content_type check + using
response.text() + json.loads
- Some plugins openapi specifications don't have operation-id so tool
name becomes "", I now added a fallback to endpoint.path if that is the
case.
- Some plugins do not have the full openapi spec URL specified in their
description, so parsing of it fails. This is now fixed with the
detection of such cases and than merging plugins config url + openapi
spec path.
- If users add a bunch of plugins, there was no height or scrolling
inside of a PluginChooser dropdown, this is fixed now. (will leave this
out as #3050 already solves this)

---------

Co-authored-by: Oliver <olivergestanley@gmail.com>