Add POST /api/v1/admin/merge_users (#3475)
First backend support for #3246 : Adds a new `/api/v1/admin/merge_users`
endpoint to merge one or more source user accounts into an existing
destination user account. The source user accounts are deleted in the
process and all objects that belonged to the source users are
transferred to the destination user. Rows belonging to the source
accounts in the `user_stats` and `troll_stats` tables are also deleted.