3517 - Send reports to discord via celery task (#3566)
It should fix: https://github.com/LAION-AI/Open-Assistant/issues/3517

In debeb353cc0104a0a330637d078b9b0bc41afe84 we reverted the original
implementation of the task as it broke the production environment. The
problem was that celery tasks need objects that could be JSON
serializable and we were passing a `Message` instance, so just passing
the details fixed the issue.

![Screenshot from 2023-07-13
08-04-49](https://github.com/LAION-AI/Open-Assistant/assets/33068707/1e76d71f-7182-49ec-8e94-38ef2fce56a2)

A few things that I noticed:
- There's no `pre-commit` version pinned in the requirements.txt so I
just used the latest one (3.3.3). In the past I had experienced issues
with this so it might be a good idea to pin it.
- There are a few automated tests, I was about to write them but found
no installation of `pytest` or similar, if you consider so, I'm happy to
write them and install the necessary libraries as eventually that will
make the codebase more robust.