Add Wikipedia Persian Dataset (#3629)
Currently, the Open-assistant model doesn't support Farsi. This is a
text-only dataset to learn Farsi (Persian).

One of my friends fine-tuned LLaMa on this dataset and It could
understand Farsi grammar and word usage very well. If the Open-assistant
team wants to add support to Farsi, this should be the first step.

I have transformed the dataset into the standard that has been mentioned
[here](https://projects.laion.ai/Open-Assistant/docs/data/datasets) and
uploaded it to [my huggingface
account](https://huggingface.co/datasets/pourmand1376/fa-wikipedia).


- #2974